refactor: Extract authentication routes and logic into authController and authRoutes
