+++ lib/handlers.js
@@ -8,46 +8,9 @@ exports.api = {}
 exports.home = (req, res) => {
     res.render('home')
 }
-exports.cadastro = (req, res) => {
-    if(req.session.user){
-        res.redirect('/')
-    }else{
-        res.render('cadastro')
-    }
-}
 
 exports.cadastrarPratos = (req, res) => {
     res.render('novosPratos',{ user: req.session.user, rota: req.params.username} )
-}
-exports.login = (req, res) => {
-    if(req.session.user){
-        res.redirect('/')
-    }else{
-        res.render('login')
-    }
-}
-exports.api.login = async(req, res) => {
-    let username = req.body.username
-    let password = req.body.password
-    try{
-        let stmt = db.prepare('SELECT * FROM users WHERE username = ?')
-        const user = stmt.get(username)
-        if(!user){
-            return res.status(401).json({err: 'Username ou senha errada'})
-        }
-        const match = await bcrypt.compare(password, user.senha)
-        if(!match){
-            return res.status(401).json({err: 'Username ou senha errada'})
-        }
-        req.session.user = {
-            id: user.id, 
-            username: username
-        };
-        return res.status(201).json({ redirect: '/', mensagem: 'Deu certo' })
-    }catch(err){
-        console.log('err')
-        return res.status(500).json({err: 'Erro interno no servidor'})
-    }
 }
 exports.perfil = (req, res) => {
     let id = req.session?.user?.id || null
@@ -143,26 +106,6 @@ exports.restaurantes = (req, res) => {
         FROM users`)
     let result = stmt.all()
     res.render('restaurantes', { result })
-}
-exports.api.cadastro = async(req, res) => {
-    const username = req.body.username
-    const pass = req.body.password
-    const name = req.body.name
-    const row = db.prepare('SELECT 1 FROM users WHERE username = ?').get(username)
-    const image = 'default.png' //aparentemente a p#### do sql nao tem como alterar uma tabela ja existente colocando o valor default em uma coluna, entao la vamos nós fazer isso igual os íncas e estou com preguiça de recriar a tabela e copiar os dados
-        if(row){
-            res.status(500).json({err: 'nome ja existente'})
-        }
-    const hash = await bcrypt.hash(pass, 10)
-    try{
-        const stmt = db.prepare(`
-            INSERT INTO users (nome, username, senha, image_url) VALUES(?, ?, ?, ?)`)
-        stmt.run(name, username, hash, image)
-    }catch(err){
-        console.log('Erro ao salvar no banco:', err)
-        res.status(500).json({error: 'Erro interno no servidor'})
-    }
-    res.status(200).json({message: 'ok', redirect: '/login'})
 }
 exports.api.usersVerify = (req, res) => {
     const username = req.query.username
